refactor: move EvaluationCalendar under Evaluation folder
Moved the EvaluationCalendar folder into the Evaluation directory
to keep all evaluation-related modules organized in one place.
Updated all import paths accordingly.
