+++ src/Evaluation/index.ts
@@ -1,0 +1,4 @@
+export { default as CalendarView } from "./EvaluationCalendar/CalendarView"
+export { IEvaluationService, EvaluationService } from './EvaluationService';
+export { EvaluationType, Evaluation } from './EvaluationService';
+export { saveOrUpdateEvaluation } from './EvaluationService/saveOrUpdateEvaluationService';
